feat: Rewrite the read_file function for usability
This rewrites the read_file function to read a single file instead of
multiple files at a time like it was trying to do. This allows for more
control over the file being read and also handles the case where there
are multiple different pressures and temperature ranges.

Additionally this adds a new function read_all_files for reading all
files within a directory, removing the requirement for looping through a
collection of files.