nb: Update 03_Supervised_Classification
This is an update to the supervised classification notebook. Malting use
of the updated dataset and utility functions. Of note here is the
significantly improved accuracy of the learning methods, getting 98%
accuracy instead of the 95% previously. This is an excellent result
although I don't completely understand why it is the case. If it just
happens to be the larger dataset then that wouldn't be particularly
great.